pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autocorrect_linux"
version = "0.4.1"
description = "System-wide autocorrect for Linux"
readme = "README.md"
requires-python = ">=3.9, <3.12"
license = "GPL-3.0" 
authors = [
    {name = "simon0302010", email = "simon0302010@gmail.com"}
]
dependencies = [
    "click",
    "pynput",
    "requests",
    "customtkinter",
    "distance",
    "tqdm",
    "keras",
    "tensorflow",
    "numpy",
    "python-xlib"
]
classifiers = [
    "Operating System :: POSIX :: Linux"
]

[project.urls]
Homepage = "https://github.com/simon0302010/autocorrect-linux"